requirements.txt
amqp==2.5.2
asn1crypto==1.2.0
bcrypt==3.1.7
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
configparser==3.8.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
cx-Oracle==7.2.2
Django==2.2.5
django-bootstrap3==11.1.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.1.0
django-filter==2.2.0
django-rest-swagger==2.2.0
django-timezone-field==3.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.6
MarkupSafe==1.1.1
more-itertools==7.2.0
mysqlclient==1.4.4
openapi-codec==1.3.2
paramiko==2.6.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.2
redis==3.3.11
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
